chore: use uint256 for request ids

+++ src/interfaces/ISignatureReceiver.sol
@@ -9,5 +9,5 @@ interface ISignatureReceiver {
     /// @dev This function is intended to be called to provide a signature for the given `requestID`.
     /// @param requestID The unique identifier of the request associated with the signature.
     /// @param signature The cryptographic signature of the message, provided as a byte array.
-    function receiveSignature(uint64 requestID, bytes calldata signature) external;
+    function receiveSignature(uint256 requestID, bytes calldata signature) external;
 }
